requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
azure-storage-blob==12.26.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
docx2pdf==0.1.8
Flask==3.1.1
fpdf==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.3.1
openai==1.99.0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
pywin32==311
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
